+++ server/src/inspector/analysisResolver.ts
