+++ server/src/inspector/fileUtils.ts
