+++ server/src/inspector/inspector.ts
